refactor: 재구조화를 통해 프로젝트 디렉토리 구조 변경
프로젝트 구조를 DDD 원칙에 따라 재정비하기 위해
도메인 중심 디렉토리 구조로 변경.
