+++ src/main/java/com/backend/komeet/chat/repositories/ChatQRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.backend.komeet.chat.repositories;
 
-import com.backend.komeet.base.model.dtos.ChatDto;
+import com.backend.komeet.chat.model.dtos.ChatDto;
 import com.backend.komeet.chat.model.entities.QChat;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.Predicate;
